% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_classification_as_factor.R
\name{add_effect_classification}
\alias{add_effect_classification}
\title{Add effect classifications to a dataframe by comparing the confidence
intervals with a reference and thresholds}
\usage{
add_effect_classification(
  df,
  cl_columns,
  threshold,
  reference = 0,
  coarse = TRUE
)
}
\arguments{
\item{df}{A dataframe containing summary data of confidence limits. Two
columns are required containing lower and upper limits indicated by the
\code{cl_columns} argument. Any other columns are optional.}

\item{cl_columns}{A vector of 2 column names in \code{df} indicating respectively
the lower and upper confidence limits (e.g. \code{c("lcl", "ucl")}).}

\item{threshold}{A vector of either 1 or 2 thresholds. A single threshold
will be transformed into \code{reference + c(-abs(threshold), abs(threshold))}.}

\item{reference}{The null hypothesis value to compare confidence intervals
against. Defaults to 0.}

\item{coarse}{Logical, defaults to \code{TRUE}. If \code{TRUE}, add a coarse
classification to the dataframe.}
}
\value{
The returned value is a modified version of the original input
dataframe \code{df} with additional columns \code{effect_code} and \code{effect} containing
respectively the effect symbols and descriptions as ordered factor variables.
In case of \code{coarse = TRUE} (by default) also \code{effect_code_coarse} and
\code{effect_coarse} containing the coarse classification effects.
}
\description{
This function adds classified effects to a dataframe as ordered factor
variables by comparing the confidence intervals with a reference and
thresholds.
}
\details{
This function is a wrapper around \code{effectclass::classify()} and
\code{effectclass::coarse_classification()} from the \pkg{effectclass} package
(Onkelinx, 2023).

...
}
\examples{
# example code

}
\references{
Onkelinx, T. (2023). effectclass: Classification and visualisation of effects
[Computer software]. \url{https://effectclass.netlify.com}
}
