% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_ci.R
\name{basic_ci}
\alias{basic_ci}
\title{Basic confidence interval (helper)}
\usage{
basic_ci(t0, t, conf = 0.95, h = function(t) t, hinv = function(t) t)
}
\arguments{
\item{t0}{Original statistic.}

\item{t}{Numeric vector of bootstrap replicates.}

\item{conf}{Confidence level.}

\item{h}{Transformation function.}

\item{hinv}{Inverse transformation function.}
}
\value{
A matrix with four columns:
\describe{
\item{conf}{confidence level}
\item{rk_lower}{rank of lower endpoint (interpolated)}
\item{rk_upper}{rank of upper endpoint (interpolated)}
\item{ll}{lower confidence limit}
\item{ul}{upper confidence limit}
}
}
\description{
Basic confidence interval (helper)
}
\note{
This function is adapted from the internal function \code{basic.ci()}
in the \pkg{boot} package (Davison & Ripley, 1999).
}
\references{
Canty, A., & Ripley, B. (1999). boot: Bootstrap Functions (Originally by
Angelo Canty for S) [Computer software].
\url{https://CRAN.R-project.org/package=boot}

Davison, A. C., & Hinkley, D. V. (1997). Bootstrap Methods and their
Application (1st ed.). Cambridge University Press.
\doi{10.1017/CBO9780511802843}
}
\seealso{
Other interval_calculation: 
\code{\link{bca_ci}()},
\code{\link{norm_ci}()},
\code{\link{perc_ci}()}
}
\concept{interval_calculation}
