% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_cube.R
\name{bootstrap_cube}
\alias{bootstrap_cube}
\title{Perform bootstrapping over a data cube for a calculated statistic}
\usage{
bootstrap_cube(
  data_cube,
  fun,
  grouping_var,
  samples = 1000,
  ref_group = NA,
  seed = NA,
  progress = FALSE
)
}
\arguments{
\item{data_cube}{A data cube object (class 'processed_cube' or 'sim_cube',
see \code{b3gbi::process_cube()}) or a dataframe (from \verb{$data} slot of
'processed_cube' or 'sim_cube').}

\item{fun}{A function which, when applied to \code{data_cube} returns the
statistic(s) of interest. This function must return a dataframe with a column
\code{diversity_val} containing the statistic of interest.}

\item{grouping_var}{A string specifying the grouping variable(s) for the
bootstrap analysis. The output of \code{fun(data_cube)} returns a row per group.}

\item{samples}{The number of bootstrap replicates. A single positive integer.
Default is 1000.}

\item{ref_group}{A string indicating the reference group to compare the
statistic with. Default is \code{NA}, meaning no reference group is used.}

\item{seed}{A positive numeric value setting the seed for random number
generation to ensure reproducibility. If \code{NA} (default), then \code{set.seed()}
is not called at all. If not \code{NA}, then the random number generator state is
reset (to the state before calling this function) upon exiting this function.}

\item{progress}{Logical. Whether to show a progress bar. Set to \code{TRUE} to
display a progress bar, \code{FALSE} (default) to suppress it.}
}
\value{
A dataframe containing the bootstrap results with the following
columns:
\itemize{
\item \code{sample}: Sample ID of the bootstrap replicate
\item \code{est_original}: The statistic based on the full dataset per group
\item \code{rep_boot}: The statistic based on a bootstrapped dataset (bootstrap
replicate)
\item \code{est_boot}: The bootstrap estimate (mean of bootstrap replicates per
group)
\item \code{se_boot}: The standard error of the bootstrap estimate (standard
deviation of the bootstrap replicates per group)
\item \code{bias_boot}: The bias of the bootstrap estimate per group
}
}
\description{
This function generate \code{samples} bootstrap replicates of a statistic applied
to a data cube. It resamples the data cube and computes a statistic \code{fun} for
each bootstrap replicate, optionally comparing the results to a reference
group (\code{ref_group}).
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Get example data
# install.packages("devtools")
# devtools::install_github("b-cubed-eu/b3gbi")
library(b3gbi)
cube_path <- system.file(
  "extdata", "denmark_mammals_cube_eqdgc.csv",
  package = "b3gbi")
denmark_cube <- process_cube(
  cube_path,
  first_year = 2014,
  last_year = 2020)

# Function to calculate statistic of interest
# Mean observations per year
mean_obs <- function(data) {
  out_df <- aggregate(obs ~ year, data, mean) # Calculate mean obs per year
  names(out_df) <- c("year", "diversity_val") # Rename columns
  return(out_df)
}
mean_obs(denmark_cube$data)

# Perform bootstrapping
bootstrap_mean_obs <- bootstrap_cube(
  data_cube = denmark_cube$data,
  fun = mean_obs,
  grouping_var = "year",
  samples = 1000,
  seed = 123,
  progress = FALSE)
head(bootstrap_mean_obs)
\dontshow{\}) # examplesIf}
}
