% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_acceleration.R
\name{calculate_acceleration}
\alias{calculate_acceleration}
\title{Calculate acceleration}
\usage{
calculate_acceleration(
  bootstrap_samples_df,
  data_cube,
  fun,
  ...,
  grouping_var,
  ref_group = NA,
  jackknife = "usual",
  progress = FALSE
)
}
\description{
Calculate acceleration
}
\details{
Acceleration quantifies how sensitive the variability of the statistic is
to changes in the data.
\itemize{
\item \eqn{a=0}: The statistic's variability does not depend on the data
(e.g., symmetric distribution)
\item \eqn{a>0}: Small changes in the data have a large effect on the
statistic's variability (e.g., positive skew)
\item \eqn{a<0}: Small changes in the data have a smaller effect on the
statistic's variability (e.g., negative skew).
}

The acceleration term is calculated as follows:

\deqn{\hat{a} = \frac{1}{6} \frac{\sum_{i = 1}^{n}(I_i^3)}{\left( \sum_{i = 1}^{n}(I_i^2) \right)^{3/2}}}

where \eqn{I_i} denotes the influence of data point \eqn{x_i} on the
estimation of \eqn{\theta}. \eqn{I_i} can be estimated using jackknifing.
Examples are (1) the negative jackknife:
\eqn{I_i = (n-1)(\hat{\theta} - \hat{\theta}_{-i})}, and (2) the positive
jackknife \eqn{I_i = (n+1)(\hat{\theta}_{-i} - \hat{\theta})}
(Frangos & Schucany, 1990). Here, \eqn{\hat{\theta}_{-i}} is the estimated
value leaving out the \eqn{i}’th data point \eqn{x_i}. The \pkg{boot}
package also offers infinitesimal jackknife and regression estimation.
Implementation of these jackknife algorithms can be explored in the
future.
}
\references{
Canty, A., & Ripley, B. (1999). boot: Bootstrap Functions (Originally by
Angelo Canty for S) [Computer software].
\url{https://CRAN.R-project.org/package=boot}

Davison, A. C., & Hinkley, D. V. (1997). Bootstrap Methods and their
Application (1st ed.). Cambridge University Press.
\doi{10.1017/CBO9780511802843}

Frangos, C. C., & Schucany, W. R. (1990). Jackknife estimation of the
bootstrap acceleration constant. Computational Statistics & Data Analysis,
9(3), 271–281. \doi{10.1016/0167-9473(90)90109-U}
}
\seealso{
Other indicator_uncertainty: 
\code{\link{add_effect_classification}()},
\code{\link{bootstrap_cube}()},
\code{\link{calculate_bootstrap_ci}()}
}
\concept{indicator_uncertainty}
