% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_diagnostics_methods.R
\name{print.summary_cube_diagnostics}
\alias{print.summary_cube_diagnostics}
\title{Print summary diagnostics}
\usage{
\method{print}{summary_cube_diagnostics}(x, ...)
}
\description{
Displays a human-readable summary of data cube diagnostics.
}
