% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_diagnostics_rules.R
\name{temporal_coverage_rule}
\alias{temporal_coverage_rule}
\title{Temporal coverage diagnostic rule}
\usage{
temporal_coverage_rule()
}
\value{
A list describing the temporal coverage diagnostic rule.
}
\description{
Creates a diagnostic rule that evaluates the temporal coverage of a
\code{processed_cube}. The rule calculates the number of unique years present
in the cube and compares it to a threshold to determine the severity
level.
}
