% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_diagnostics_class.R
\name{validate_cube_diagnostics}
\alias{validate_cube_diagnostics}
\title{Validate cube_diagnostics structure}
\usage{
validate_cube_diagnostics(x)
}
\arguments{
\item{x}{A data frame containing diagnostic results.}
}
\value{
The validated object \code{x}. An error is thrown if required columns are
missing.
}
\description{
Internal validation function for \code{cube_diagnostics} objects. Ensures that the
required columns are present in the diagnostics table.
}
