% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_rule_class.R
\name{validate_cube_rule}
\alias{validate_cube_rule}
\title{Validate cube_rule structure}
\usage{
validate_cube_rule(x)
}
\arguments{
\item{x}{A list describing a diagnostic rule.}
}
\value{
The validated rule object \code{x}.
}
\description{
Internal validation function for \code{cube_rule} objects. Ensures that the rule
definition contains the required elements.
}
