% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_ci.R
\name{basic_ci}
\alias{basic_ci}
\title{Calculate basic bootstrap confidence interval}
\usage{
basic_ci(t0, t, conf = 0.95, h = function(t) t, hinv = function(t) t)
}
\arguments{
\item{t0}{Original statistic.}

\item{t}{Numeric vector of bootstrap replicates.}

\item{conf}{A numeric value specifying the confidence level of the interval.
Default is \code{0.95} (95 \% confidence level).}

\item{h}{A function defining a transformation. The intervals are calculated
on the scale of \code{h(t)} and the inverse function \code{hinv} applied to the
resulting intervals. It must be a function of one variable only. The default
is the identity function.}

\item{hinv}{A function, like \code{h}, which returns the inverse of \code{h}. It is
used to transform the intervals calculated on the scale of \code{h(t)} back to the
original scale. The default is the identity function. If \code{h} is supplied but
\code{hinv} is not, then the intervals returned will be on the transformed scale.}
}
\value{
A matrix with four columns:
\itemize{
\item \code{conf}: confidence level
\item \code{rk_lower}: rank of lower endpoint (interpolated)
\item \code{rk_upper}: rank of upper endpoint (interpolated)
\item \code{ll}: lower confidence limit
\item \code{ul}: lower confidence limit
}
}
\description{
This function calculates a basic confidence interval from a bootstrap sample.
It is used by \code{calculate_bootstrap_ci()}.
}
\details{
\deqn{CI_{basic} = \left[ 2\hat{\theta} - \hat{\theta}^*_{(1-\alpha/2)},
2\hat{\theta} - \hat{\theta}^*_{(\alpha/2)} \right]}

where \eqn{\hat{\theta}^*_{(\alpha/2)}} and
\eqn{\hat{\theta}^*_{(1-\alpha/2)}} are the \eqn{\alpha/2} and
\eqn{1-\alpha/2} percentiles of the bootstrap distribution, respectively.
}
\note{
This function is adapted from the internal function \code{basic.ci()}
in the \pkg{boot} package (Canty & Ripley, 1999).
}
\examples{
set.seed(123)
boot_reps <- rnorm(1000)
t0 <- mean(boot_reps)

# Basic bootstrap CI
basic_ci(t0, boot_reps, conf = 0.95)
}
\references{
Canty, A., & Ripley, B. (1999). boot: Bootstrap Functions (Originally by
Angelo Canty for S) [Computer software].
\url{https://CRAN.R-project.org/package=boot}

Davison, A. C., & Hinkley, D. V. (1997). Bootstrap Methods and their
Application (1st ed.). Cambridge University Press.
\doi{10.1017/CBO9780511802843}
}
\seealso{
Other interval_calculation: 
\code{\link{bca_ci}()},
\code{\link{norm_ci}()},
\code{\link{perc_ci}()}
}
\concept{interval_calculation}
