% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bca_ci.R
\name{bca_ci}
\alias{bca_ci}
\title{Bias-Corrected and Accelerated (BCa) confidence interval (helper)}
\usage{
bca_ci(t0, t, a, conf = 0.95, h = function(t) t, hinv = function(t) t)
}
\arguments{
\item{t0}{Original statistic.}

\item{t}{Numeric vector of bootstrap replicates.}

\item{a}{Acceleration constant.}

\item{conf}{Confidence level.}

\item{h}{Transformation function.}

\item{hinv}{Inverse transformation function.}
}
\value{
A matrix with four columns:
\describe{
\item{conf}{confidence level}
\item{rk_lower}{rank of lower endpoint (interpolated)}
\item{rk_upper}{rank of upper endpoint (interpolated)}
\item{ll}{lower confidence limit}
\item{ul}{upper confidence limit}
}
}
\description{
Bias-Corrected and Accelerated (BCa) confidence interval (helper)
}
\note{
This function is adapted from the internal function \code{bca.ci()}
in the \pkg{boot} package (Davison & Ripley, R Core Team).
Credit: \pkg{boot} authors (A. C. Davison, B. D. Ripley, and R Core Team).
Licensed under the same terms as R itself.
}
\references{
Canty, A., & Ripley, B. (1999). boot: Bootstrap Functions (Originally by
Angelo Canty for S) [Computer software].
\url{https://CRAN.R-project.org/package=boot}

Davison, A. C., & Hinkley, D. V. (1997). Bootstrap Methods and their
Application (1st ed.). Cambridge University Press.
\doi{10.1017/CBO9780511802843}
}
\seealso{
Other interval_calculation: 
\code{\link{basic_ci}()},
\code{\link{norm_ci}()},
\code{\link{perc_ci}()}
}
\concept{interval_calculation}
