% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_acceleration.R
\name{calculate_acceleration}
\alias{calculate_acceleration}
\title{Calculate acceleration for a dataframe with bootstrap replicates}
\usage{
calculate_acceleration(
  bootstrap_samples_df,
  data_cube,
  fun,
  ...,
  grouping_var,
  ref_group = NA,
  jackknife = "usual",
  progress = FALSE
)
}
\arguments{
\item{bootstrap_samples_df}{A dataframe containing the bootstrap replicates,
where each row represents a bootstrap sample. As returned by
\code{bootstrap_cube()}. Apart from the \code{grouping_var} column, the following
columns should be present:
\itemize{
\item \code{est_original}: The statistic based on the full dataset per group
}}

\item{data_cube}{A data cube object (class
'processed_cube' or 'sim_cube', see \code{b3gbi::process_cube()}) or a dataframe
(from \verb{$data} slot of 'processed_cube' or 'sim_cube'). As used by
\code{bootstrap_cube()}. To limit runtime, we recommend using a
dataframe with custom function as \code{fun}.}

\item{fun}{A function which, when applied to
\code{data_cube} returns the statistic(s) of interest. This function must return a
dataframe with a column \code{diversity_val} containing the statistic of interest.
As used by \code{bootstrap_cube()}.}

\item{...}{Additional arguments passed on to \code{fun}.}

\item{grouping_var}{A character vector specifying the grouping variable(s)
for the bootstrap analysis. The function \code{fun(data_cube, ...)} should return
a row per group. The specified variables must not be redundant, meaning they
should not contain the same information (e.g., \code{"time_point"} (1, 2, 3) and
\code{"year"} (2000, 2001, 2002) should not be used together if \code{"time_point"} is
just an alternative encoding of \code{"year"}).
This variable is used to split the dataset into groups for separate
acceleration calculations.}

\item{ref_group}{A string indicating the
reference group to compare the statistic with. Default is \code{NA}, meaning no
reference group is used.
As used by \code{bootstrap_cube()}.}

\item{jackknife}{A string specifying the
jackknife resampling method for calculating the influence values.
\itemize{
\item \code{"usual"}: Negative jackknife (default if BCa is selected).
\item \code{"pos"}: Positive jackknife
}}

\item{progress}{Logical. Whether to show a progress bar for jackknifing. Set
to \code{TRUE} to display a progress bar, \code{FALSE} (default) to suppress it.}
}
\value{
A dataframe containing the acceleration values per \code{grouping_var}.
}
\description{
This function calculates acceleration values for a dataframe containing
bootstrap replicates. Acceleration quantifies how sensitive the variability
of a statistic is to changes in the data. It can be used for BCa interval
calculation \code{calculate_bootstrap_ci()}.
}
\details{
Acceleration quantifies how sensitive the variability of a statistic is
to changes in the data.
\itemize{
\item \eqn{a=0}: The statistic's variability does not depend on the data
(e.g., symmetric distribution)
\item \eqn{a>0}: Small changes in the data have a large effect on the
statistic's variability (e.g., positive skew)
\item \eqn{a<0}: Small changes in the data have a smaller effect on the
statistic's variability (e.g., negative skew).
}

The acceleration term is calculated as follows (Davison & Hinkley, 1997,
Chapter 5; see also the \pkg{boot} package in R (Canty & Ripley, 1999)):

\deqn{\hat{a} = \frac{1}{6} \frac{\sum_{i = 1}^{n}(I_i^3)}{\left( \sum_{i = 1}^{n}(I_i^2) \right)^{3/2}}}

where \eqn{I_i} denotes the influence of data point \eqn{x_i} on the
estimation of \eqn{\theta}. \eqn{I_i} can be estimated using jackknifing.
Examples are (1) the negative jackknife:
\eqn{I_i = (n-1)(\hat{\theta} - \hat{\theta}_{-i})}, and (2) the positive
jackknife \eqn{I_i = (n+1)(\hat{\theta}_{-i} - \hat{\theta})}
(Frangos & Schucany, 1990). Here, \eqn{\hat{\theta}_{-i}} is the estimated
value leaving out the \eqn{i}’th data point \eqn{x_i}. The \pkg{boot}
package also offers infinitesimal jackknife and regression estimation.
Implementation of these jackknife algorithms can be explored in the
future.
}
\examples{
# Get example data
# install.packages("b3gbi", repos = "https://b-cubed-eu.r-universe.dev")
library(b3gbi)
cube_path <- system.file(
  "extdata", "denmark_mammals_cube_eqdgc.csv",
  package = "b3gbi")
denmark_cube <- process_cube(
  cube_path,
  first_year = 2014,
  last_year = 2020)

# Function to calculate statistic of interest
# Mean observations per year
mean_obs <- function(data) {
  out_df <- aggregate(obs ~ year, data, mean) # Calculate mean obs per year
  names(out_df) <- c("year", "diversity_val") # Rename columns
  return(out_df)
}
mean_obs(denmark_cube$data)

# Perform bootstrapping
\donttest{
bootstrap_mean_obs <- bootstrap_cube(
  data_cube = denmark_cube$data,
  fun = mean_obs,
  grouping_var = "year",
  samples = 1000,
  seed = 123,
  progress = FALSE)
head(bootstrap_mean_obs)

# Calculate acceleration
acceleration_df <- calculate_acceleration(
  bootstrap_samples_df = bootstrap_mean_obs,
  data_cube = denmark_cube$data,
  fun = mean_obs,
  grouping_var = "year",
  progress = FALSE)
acceleration_df
}
}
\references{
Canty, A., & Ripley, B. (1999). boot: Bootstrap Functions (Originally by
Angelo Canty for S) [Computer software].
\url{https://CRAN.R-project.org/package=boot}

Davison, A. C., & Hinkley, D. V. (1997). Bootstrap Methods and their
Application (1st ed.). Cambridge University Press.
\doi{10.1017/CBO9780511802843}

Frangos, C. C., & Schucany, W. R. (1990). Jackknife estimation of the
bootstrap acceleration constant. Computational Statistics & Data Analysis,
9(3), 271–281. \doi{10.1016/0167-9473(90)90109-U}
}
\seealso{
Other indicator_uncertainty: 
\code{\link{add_effect_classification}()},
\code{\link{bootstrap_cube}()},
\code{\link{calculate_bootstrap_ci}()}
}
\concept{indicator_uncertainty}
