% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_diagnostics_methods.R
\name{print.summary_cube_diagnostics}
\alias{print.summary_cube_diagnostics}
\title{Print summary diagnostics}
\usage{
\method{print}{summary_cube_diagnostics}(x, ...)
}
\arguments{
\item{x}{A \code{summary_cube_diagnostics} object returned by
\code{summary(diagnose_cube(...))}.}

\item{...}{Additional arguments passed to other methods (currently unused).}
}
\description{
Displays a human-readable summary of data cube diagnostics.
}
