% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_diagnostics_rules.R
\name{rule_obs_min_total}
\alias{rule_obs_min_total}
\title{Minimum total number of observations diagnostic rule}
\usage{
rule_obs_min_total()
}
\value{
An object of class \code{cube_rule}.
}
\description{
Creates a diagnostic rule that evaluates whether a data cube contains a
sufficient number of total observations. The rule sums
the number of observations present in the cube and compares it to a
threshold to determine the severity level.
}
